requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.13.4
    # via libpysal
catboost==1.2.8
    # via -r requirements.in
certifi==2025.6.15
    # via
    #   pyogrio
    #   pyproj
    #   requests
charset-normalizer==3.4.2
    # via requests
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.58.4
    # via matplotlib
geopandas==1.1.0
    # via
    #   -r requirements.in
    #   libpysal
graphviz==0.21
    # via catboost
idna==3.10
    # via requests
imbalanced-learn==0.13.0
    # via -r requirements.in
joblib==1.5.1
    # via
    #   imbalanced-learn
    #   scikit-learn
kiwisolver==1.4.8
    # via matplotlib
libpysal==4.13.0
    # via -r requirements.in
lightgbm==4.6.0
    # via -r requirements.in
matplotlib==3.10.3
    # via
    #   -r requirements.in
    #   catboost
narwhals==1.43.1
    # via plotly
numpy==2.3.0
    # via
    #   catboost
    #   contourpy
    #   geopandas
    #   imbalanced-learn
    #   libpysal
    #   lightgbm
    #   matplotlib
    #   pandas
    #   pyogrio
    #   scikit-learn
    #   scipy
    #   shapely
    #   xgboost
packaging==25.0
    # via
    #   geopandas
    #   libpysal
    #   matplotlib
    #   plotly
    #   pyogrio
pandas==2.3.0
    # via
    #   -r requirements.in
    #   catboost
    #   geopandas
    #   libpysal
pillow==11.2.1
    # via matplotlib
platformdirs==4.3.8
    # via libpysal
plotly==6.1.2
    # via catboost
pyogrio==0.11.0
    # via geopandas
pyparsing==3.2.3
    # via matplotlib
pyproj==3.7.1
    # via geopandas
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
requests==2.32.4
    # via libpysal
scikit-learn==1.6.1
    # via
    #   -r requirements.in
    #   imbalanced-learn
    #   libpysal
    #   sklearn-compat
scipy==1.15.3
    # via
    #   catboost
    #   imbalanced-learn
    #   libpysal
    #   lightgbm
    #   scikit-learn
    #   xgboost
shapely==2.1.1
    # via
    #   geopandas
    #   libpysal
six==1.17.0
    # via
    #   catboost
    #   python-dateutil
sklearn-compat==0.1.3
    # via imbalanced-learn
soupsieve==2.7
    # via beautifulsoup4
threadpoolctl==3.6.0
    # via
    #   imbalanced-learn
    #   scikit-learn
typing-extensions==4.14.0
    # via beautifulsoup4
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
xgboost==3.0.2
    # via -r requirements.in
